+++ CHANGES.md
@@ -4,4 +4,5 @@
 
 **Features**
 
-- Currency formatting filter.
+- New currency formatting filter.
+- New convenience "money" filters. `money`, `money_with_currency`, `money_without_currency` and `money_without_trailing_zeros` behave similarly to Shopify's equivalents.